+++ processor/src/main/java/arez/processor/Constants.java
@@ -36,6 +36,8 @@ final class Constants
   static final String SCOPE_ANNOTATION_CLASSNAME = "javax.inject.Scope";
   static final String DAGGER_MODULE_CLASSNAME = "dagger.Module";
   static final String NONNULL_ANNOTATION_CLASSNAME = "javax.annotation.Nonnull";
+  static final String NULLABLE_ANNOTATION_CLASSNAME = "javax.annotation.Nullable";
+  static final String DEPRECATED_ANNOTATION_CLASSNAME = "java.lang.Deprecated";
 
   private Constants()
   {